+++ src/plone/restapi/batching.py
@@ -1,7 +1,7 @@
 from plone.batching.batch import Batch
 from plone.restapi.deserializer import json_body
-from six.moves.urllib.parse import parse_qsl
-from six.moves.urllib.parse import urlencode
+from urllib.parse import parse_qsl
+from urllib.parse import urlencode
 
 
 DEFAULT_BATCH_SIZE = 25