+++ src/plone/restapi/demo/demo.py
@@ -3,7 +3,7 @@ from Products.Five.browser import BrowserView  # pragma: no cover
 
 class InternalServerErrorView(BrowserView):  # pragma: no cover
     def __call__(self):
-        from six.moves.urllib.error import HTTPError
+        from urllib.error import HTTPError
 
         raise HTTPError(
             "http://nohost/plone/internal_server_error",