+++ src/plone/restapi/services/content/add.py
@@ -1,5 +1,8 @@
+from AccessControl import getSecurityManager
 from Acquisition import aq_base
 from Acquisition.interfaces import IAcquirer
+from plone.app.multilingual.interfaces import IPloneAppMultilingualInstalled
+from plone.app.multilingual.interfaces import ITranslationManager
 from plone.restapi.deserializer import json_body
 from plone.restapi.exceptions import DeserializationError
 from plone.restapi.interfaces import IDeserializeFromJson
@@ -7,7 +10,8 @@ from plone.restapi.interfaces import ISerializeToJson
 from plone.restapi.services import Service
 from plone.restapi.services.content.utils import add
 from plone.restapi.services.content.utils import create
-from Products.CMFPlone.utils import getFSVersionTuple
+from Products.CMFCore.permissions import ManagePortal
+from Products.CMFCore.utils import getToolByName
 from Products.CMFPlone.utils import safe_hasattr
 from zExceptions import BadRequest
 from zExceptions import Unauthorized
@@ -16,20 +20,8 @@ from zope.event import notify
 from zope.interface import alsoProvides
 from zope.lifecycleevent import ObjectCreatedEvent
 from zope.component import getMultiAdapter
-from Products.CMFCore.utils import getToolByName
 
 import plone.protect.interfaces
-import pkg_resources
-import six
-
-
-try:
-    pkg_resources.get_distribution("plone.app.multilingual")
-    PAM_INSTALLED = True
-except pkg_resources.DistributionNotFound:
-    PAM_INSTALLED = False
-
-PLONE5 = getFSVersionTuple()[0] >= 5
 
 
 class FolderPost(Service):
@@ -43,6 +35,7 @@ class FolderPost(Service):
         title = data.get("title", None)
         translation_of = data.get("translation_of", None)
         language = data.get("language", None)
+        uid = data.get("UID", None)
 
         if not type_:
             raise BadRequest("Property '@type' is required")
@@ -50,6 +43,13 @@ class FolderPost(Service):
         # Disable CSRF protection
         if "IDisableCSRFProtection" in dir(plone.protect.interfaces):
             alsoProvides(self.request, plone.protect.interfaces.IDisableCSRFProtection)
+
+        sm = getSecurityManager()
+        # ManagePortal is required to set the uid of an object during creation
+        if uid and not sm.checkPermission(ManagePortal, self.context):
+            self.request.response.setStatus(403)
+            msg = "Setting UID of an object requires Manage Portal permission"
+            return dict(error=dict(type="Forbidden", message=msg))
 
         try:
             obj = create(self.context, type_, id_=id_, title=title)
@@ -84,27 +84,24 @@ class FolderPost(Service):
         if temporarily_wrapped:
             obj = aq_base(obj)
 
+        if uid:
+            setattr(obj, "_plone.uuid", uid)
+
         if not getattr(deserializer, "notifies_create", False):
             notify(ObjectCreatedEvent(obj))
 
         obj = add(self.context, obj, rename=not bool(id_))
 
         # Link translation given the translation_of property
-        if PAM_INSTALLED and PLONE5:
-            from plone.app.multilingual.interfaces import (
-                IPloneAppMultilingualInstalled,
-            )  # noqa
-            from plone.app.multilingual.interfaces import ITranslationManager
-
-            if (
-                IPloneAppMultilingualInstalled.providedBy(self.request)
-                and translation_of
-                and language
-            ):
-                source = self.get_object(translation_of)
-                if source:
-                    manager = ITranslationManager(source)
-                    manager.register_translation(language, obj)
+        if (
+            IPloneAppMultilingualInstalled.providedBy(self.request)
+            and translation_of
+            and language
+        ):
+            source = self.get_object(translation_of)
+            if source:
+                manager = ITranslationManager(source)
+                manager.register_translation(language, obj)
 
         self.request.response.setStatus(201)
         self.request.response.setHeader("Location", obj.absolute_url())
@@ -129,12 +126,8 @@ class FolderPost(Service):
         if key.startswith(portal.absolute_url()):
             # Resolve by URL
             key = key[len(portal.absolute_url()) + 1 :]
-            if six.PY2:
-                key = key.encode("utf8")
             return portal.restrictedTraverse(key, None)
         elif key.startswith("/"):
-            if six.PY2:
-                key = key.encode("utf8")
             # Resolve by path
             return portal.restrictedTraverse(key.lstrip("/"), None)
         else: