+++ src/plone/restapi/tests/test_services_email_notification.py
@@ -12,17 +12,6 @@ import transaction
 import unittest
 
 
-try:
-    from Products.CMFPlone.factory import _IMREALLYPLONE5  # noqa
-except ImportError:
-    PLONE5 = False
-else:
-    PLONE5 = True
-
-
-@unittest.skipIf(
-    not PLONE5, "email notification not implemented for Plone < 5."
-)  # noqa
 class EmailNotificationEndpoint(unittest.TestCase):
 
     layer = PLONE_RESTAPI_DX_FUNCTIONAL_TESTING